+++ CHANGELOG.md
@@ -5,3 +5,8 @@
 * Semantics understanding
 * Speech synthesis
 * Wakeup
+
+# 0.0.2
+
+* Fix bug that Android 11 cannot use wakeup function. The `resPath` parameter of `ivw` configuration will no longer take effect,
+  also, you need to delete the `res_path` configuration in the `vtn.ini` file.